splitting installation from execution
